% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{assert_color_options}
\alias{assert_color_options}
\title{Assert special color inputs}
\usage{
assert_color_options(color_low = NULL, color_high = NULL, viridis = NULL)
}
\arguments{
\item{color_low, color_high}{(Optional) A \code{\link[base:character]{character}}
vector representing color names or hexadecimal color codes
(Default: \code{NULL}).}

\item{viridis}{(Optional) A \code{\link[base:character]{character}} vector
representing viridis color palette names (Default: \code{NULL}).}
}
\value{
An \link[base:invisible]{invisible} \code{TRUE} if the input is valid;
an error message otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
\code{assert_color_options()} ensures that the provided color options are valid.
It checks if the \code{color_low} and \code{color_high} arguments are valid colors
and are provided together, or, alternatively, ensures that \code{viridis} color
palette names are valid. These options are mutually exclusive.
}
\examples{
assert_color_options(color_low = "#FF5733", color_high = "#33FF57")

assert_color_options(
  color_low = c("#FF5733", "#FFF"),
  color_high = c("#33FF57", "#000")
)

assert_color_options(viridis = "magma")
assert_color_options(viridis = c("magma", "viridis"))
}
