% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{assert_ggplot_label}
\alias{assert_ggplot_label}
\title{Assert the validity of a \code{ggplot2} label}
\usage{
assert_ggplot_label(x, null_ok = FALSE)
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} string or a
\code{\link[latex2exp:TeX]{latexexpression}} object.}

\item{null_ok}{(Optional) A \code{\link[base:logical]{logical}} flag indicating if
\code{NULL} values are allowed (Default: \code{FALSE}).}
}
\value{
\code{x} (as \link[base:invisible]{invisible}) if it passes the test; an
error message otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{assert_ggplot_label()} ensures that the input is a valid
\href{https://ggplot2.tidyverse.org/}{\code{ggplot2}} label.

A \code{ggplot2} label can be a \code{\link[base:character]{character}} string or a
\code{\link[latex2exp:TeX]{latexexpression}} object.
}
\examples{
assert_ggplot_label("Test label")
assert_ggplot_label(latex2exp::TeX("$\\\\overline{x}$"))
}
