% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_data}
\alias{check_data}
\alias{test_data}
\alias{assert_data}
\title{Check if a dataset exists in a package}
\usage{
check_data(data, package)

test_data(data, package)

assert_data(data, package)
}
\arguments{
\item{data}{A \code{\link[base:character]{character}} string of the dataset name.}

\item{package}{A \code{\link[base:character]{character}} string of the package name.}
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{assertion_*}: An \link[base:invisible]{invisible} \code{TRUE} if it passes the
test; an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
if (requireNamespace("datasets", quietly = TRUE)) {
  test_data("mtcars", "datasets")
  #> [1] TRUE # Expected
}
}
