% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms.R
\name{check_hms}
\alias{check_hms}
\alias{test_hms}
\alias{assert_hms}
\alias{expect_hms}
\title{Check if an argument is a \code{\link[hms:hms]{hms}} object}
\usage{
check_hms(
  x,
  lower = -Inf,
  upper = Inf,
  any_missing = TRUE,
  null_ok = FALSE,
  .names = deparse(substitute(x))
)

test_hms(
  x,
  lower = -Inf,
  upper = Inf,
  any_missing = TRUE,
  null_ok = FALSE,
  .names = deparse(substitute(x))
)

assert_hms(
  x,
  lower = -Inf,
  upper = Inf,
  any_missing = TRUE,
  null_ok = FALSE,
  .names = deparse(substitute(x)),
  .var.name = checkmate::vname(x),
  add = NULL
)

expect_hms(
  x,
  lower = -Inf,
  upper = Inf,
  any_missing = TRUE,
  null_ok = FALSE,
  .names = deparse(substitute(x)),
  info = NULL,
  label = checkmate::vname(x)
)
}
\arguments{
\item{x}{Any R object.}

\item{lower}{(Optional) A value of the same class as \code{x} indicating
the minimum value of \code{x} (Default: \code{-Inf}).}

\item{upper}{(Optional) A value of the same class as \code{x} indicating
the maximum value of \code{x} (Default: \code{Inf}).}

\item{any_missing}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
if missing values are allowed (Default: \code{TRUE}).}

\item{null_ok}{(Optional) A \code{\link[base:logical]{logical}} flag indicating if
\code{NULL} values are allowed (Default: \code{FALSE}).}

\item{.names}{(Optional) A \code{\link[base:character]{character}} vector specifying
the names of the objects being tested. This argument is used internally
and should not be set by the user.}

\item{.var.name}{(Optional) A \code{\link[base:character]{character}} string
indicating the name of the checked object to print in assertions.
This argument is used internally and should not be set by the user.}

\item{add}{(Optional) A \code{\link[checkmate:AssertCollection]{AssertCollection}}
object indicating the collection to store assertion messages. This
argument is used internally and should not be set by the user.}

\item{info}{(Optional) A \code{\link[base:character]{character}} string with extra
information to be included in the message for the `testthat`` reporter.
This argument is used internally and should not be set by the user.}

\item{label}{(Optional) A \code{\link[base:character]{character}} string
indicating the name of the checked object to print in messages.
This argument is used internally and should not be set by the user.}
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{check_*}: \code{TRUE} if it passes the test; a string with a message otherwise.
\item \verb{assertion_*}: The same input (as \link[base:invisible]{invisible}) if it
passes the test; an error message otherwise.
\item \verb{expect_*}: The same input (as \link[base:invisible]{invisible}) if it
passes the test; an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\verb{*_hms()} check if an argument is a \code{\link[lubridate:duration]{hms}} object.
}
\examples{
test_hms(1)
#> [1] FALSE # Expected

test_hms(lubridate::dhours(1))
#> [1] FALSE # Expected

test_hms(hms::parse_hm("01:00"))
#> [1] TRUE # Expected

test_hms(hms::parse_hm("01:00"), lower = hms::parse_hm("02:00"))
#> [1] FALSE # Expected

test_hms(hms::parse_hm("02:00"), upper = hms::parse_hm("01:00"))
#> [1] FALSE # Expected

test_hms(c(hms::parse_hm("01:00"), NA), any_missing = FALSE)
#> [1] FALSE # Expected

test_hms(NULL, null_ok = FALSE)
#> [1] FALSE # Expected
}
