% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{check_interactive}
\alias{check_interactive}
\alias{test_interactive}
\alias{assert_interactive}
\title{Check if R is being used interactively}
\usage{
check_interactive()

test_interactive()

assert_interactive()
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{check_*}: \code{TRUE} if it passes the test; a string with a message otherwise.
\item \verb{assertion_*}: The same input (as \link[base:invisible]{invisible}) if it
passes the test; an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
if (interactive()) {
  test_interactive()
  #> [1] TRUE # Expected
}
}
