% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.R
\name{check_internet}
\alias{check_internet}
\alias{test_internet}
\alias{assert_internet}
\title{Check for internet connectivity}
\usage{
check_internet()

test_internet()

assert_internet()
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{check_*}: \code{TRUE} if it passes the test; a string with a message otherwise.
\item \verb{assertion_*}: The same input (as \link[base:invisible]{invisible}) if it
passes the test; an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
if (interactive()) {
  test_internet()
}
}
