% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_empty.R
\name{check_not_empty}
\alias{check_not_empty}
\alias{test_not_empty}
\alias{assert_not_empty}
\title{Check if an argument has length greater than zero}
\usage{
check_not_empty(x, null_ok = FALSE, .names = deparse(substitute(x)))

test_not_empty(x, null_ok = FALSE, .names = deparse(substitute(x)))

assert_not_empty(x, null_ok = FALSE, .names = deparse(substitute(x)))
}
\arguments{
\item{x}{Any R object.}

\item{null_ok}{(Optional) A \code{\link[base:logical]{logical}} flag indicating if
\code{NULL} values are allowed (Default: \code{FALSE}).}

\item{.names}{(Optional) A \code{\link[base:character]{character}} vector specifying
the names of the objects being tested. This argument is used internally
and should not be set by the user.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\verb{*_not_empty()} check if an argument has length greater than zero.
}
\examples{
test_not_empty(character())
#> [1] FALSE # Expected

test_not_empty(1)
#> [1] TRUE # Expected
}
