% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{check_numeric}
\alias{check_numeric}
\alias{assert_numeric}
\alias{test_numeric}
\title{Check if an argument is a \code{\link[base:numeric]{numeric}} object}
\usage{
check_numeric(
  x,
  lower = -Inf,
  upper = Inf,
  finite = FALSE,
  any_missing = TRUE,
  all_missing = TRUE,
  len = NULL,
  min_len = NULL,
  max_len = NULL,
  null_ok = FALSE,
  .names = deparse(substitute(x))
)

assert_numeric(
  x,
  lower = -Inf,
  upper = Inf,
  finite = FALSE,
  any_missing = TRUE,
  all_missing = TRUE,
  len = NULL,
  min_len = NULL,
  max_len = NULL,
  null_ok = FALSE,
  .names = deparse(substitute(x))
)

test_numeric(
  x,
  lower = -Inf,
  upper = Inf,
  finite = FALSE,
  any_missing = TRUE,
  all_missing = TRUE,
  len = NULL,
  min_len = NULL,
  max_len = NULL,
  null_ok = FALSE,
  .names = deparse(substitute(x))
)
}
\arguments{
\item{x}{Any R object.}

\item{lower}{(Optional) A value of the same class as \code{x} indicating
the minimum value of \code{x} (Default: \code{-Inf}).}

\item{upper}{(Optional) A value of the same class as \code{x} indicating
the maximum value of \code{x} (Default: \code{Inf}).}

\item{finite}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
if all values must be \link[base:is.finite]{finite} (Default: \code{TRUE}).}

\item{any_missing}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
if missing values are allowed (Default: \code{TRUE}).}

\item{all_missing}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
if all missing values are allowed (Default: \code{TRUE}).}

\item{len}{(Optional) An integer number indicating the expected length of \code{x}
(Default: \code{NULL}).}

\item{min_len}{(Optional) An integer number indicating the minimum length of
\code{x} (Default: \code{NULL}).}

\item{max_len}{(Optional) An integer number indicating the maximum length of
\code{x} (Default: \code{NULL}).}

\item{null_ok}{(Optional) A \code{\link[base:logical]{logical}} flag indicating if
\code{NULL} values are allowed (Default: \code{FALSE}).}

\item{.names}{(Optional) A \code{\link[base:character]{character}} vector specifying
the names of the objects being tested. This argument is used internally
and should not be set by the user.}
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{check_*}: \code{TRUE} if it passes the test; a string with a message otherwise.
\item \verb{assertion_*}: The same input (as \link[base:invisible]{invisible}) if it
passes the test; an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
test_numeric("a")
#> [1] FALSE # Expected

test_numeric(lubridate::duration())
#> [1] FALSE # Expected

test_numeric(1:10)
#> [1] TRUE # Expected

x <- 1:2
check_numeric(x, len = 1) |> cli::cli_alert_warning()
#> ! x must have 1 element. # Expected

check_numeric(x, min_len = 3) |> cli::cli_alert_warning()
#> ! x must have 3 or more elements. # Expected

check_numeric(x, max_len = 1) |> cli::cli_alert_warning()
#> ! x must have 1 or less elements. # Expected

x <- 1
check_numeric(1, min_len = 2, max_len = 3) |> cli::cli_alert_warning()
#> ! x must have a length between 2 and 3. # Expected
}
