% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettycheck-package.R
\docType{package}
\name{prettycheck-package}
\alias{prettycheck}
\alias{prettycheck-package}
\title{prettycheck: Pretty assertive programming}
\description{
Provides a set of functions that produce pretty and informative error messages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/danielvartan/prettycheck/}
  \item Report bugs at \url{https://github.com/danielvartan/prettycheck/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
