% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identical.R
\name{test_identical}
\alias{test_identical}
\alias{check_identical}
\alias{assert_identical}
\alias{expect_identical}
\title{Check if multiple objects are identical}
\usage{
test_identical(..., type = "value")

check_identical(..., type = "value")

assert_identical(..., type = "value")

expect_identical(..., type = "value")
}
\arguments{
\item{...}{Objects to compare.}

\item{type}{A string corresponding to the type of comparison to perform.
Valid values: \code{"value"}, \code{"length"}, and \code{"class"} (default: \code{"value"}).}
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{check_*}: An \link[=invisible]{invisible} \code{TRUE} if it passes the test;
a warning message otherwise.
\item \verb{assertion_*}: An \link[=invisible]{invisible} \code{TRUE} if it passes the test;
an error message otherwise.
\item \verb{expect_*}: An \link[=invisible]{invisible} \code{TRUE} if it passes the test;
an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
x <- 1; y <- 1
test_identical(x, y, type = "value")
#> TRUE # Expected

x <- 1; y <- 2
check_identical(x, y, type = "value")
#> ! x and y must have identical values. # Expected

x <- letters; y <- LETTERS
test_identical(x, y, type = "length")
#> TRUE # Expected

x <- letters; y <- 1:2
check_identical(x, y, type = "length")
#> ! x and y must have identical lengths. # Expected

x <- "a"; y <- c("a", "b")
test_identical(x, y, type = "class")
#> TRUE # Expected

x <- "a"; y <- 1
check_identical(x, y, type = "class")
#> ! x and y must have identical classes. # Expected
}
