% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{test_length}
\alias{test_length}
\alias{check_length}
\alias{assert_length}
\title{Check if an argument has a specific length}
\usage{
test_length(x, len = NULL, min_len = NULL, max_len = NULL, null_ok = FALSE)

check_length(x, len = NULL, min_len = NULL, max_len = NULL, null_ok = FALSE)

assert_length(x, len = NULL, min_len = NULL, max_len = NULL, null_ok = FALSE)
}
\arguments{
\item{x}{An R object.}

\item{len}{(optional) an integer number indicating the expected length of \code{x}
(default: \code{NULL}).}

\item{min_len}{(optional) an integer number indicating the minimum length of
\code{x} (default: \code{NULL}).}

\item{max_len}{(optional) an integer number indicating the maximum length of
\code{x} (default: \code{NULL}).}

\item{null_ok}{(optional) a \code{\link[base:logical]{logical}} flag indicating
if \code{x} can be \code{NULL} (default: \code{FALSE}).}
}
\value{
\itemize{
\item \verb{test_*}: \code{TRUE} if it passes the test; \code{FALSE} otherwise.
\item \verb{check_*}: An \link[=invisible]{invisible} \code{TRUE} if it passes the test;
a warning message otherwise.
\item \verb{assertion_*}: An \link[=invisible]{invisible} \code{TRUE} if it passes the test;
an error message otherwise.
\item \verb{expect_*}: An \link[=invisible]{invisible} \code{TRUE} if it passes the test;
an error message otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
test_length(1, len = 1)
#> [1] TRUE # Expected
}
